Added fixes for Posix implementation
This changeset should conclude all necessary changes for the posix
implementation of threads and sempahores.

The semaphore is actually a mutex on posix platforms. It will assert
if anything other than a binary semaphore is created.

These wrappers depend on boost::function, so boost is a dependency
for these to build.